backend/requirements.txt
Flask>=3.0,<4.0
pymongo>=4.8,<5.0
python-dotenv>=1.0,<2.0
Flask-Login>=0.6,<1.0
pytest>=8.0,<9.0
